refactor(cli): extract skills + roles command families into cli/commands/skills.ts
